+++ test/loader.js
@@ -8,21 +8,30 @@ let bitdepth;
 
 // Browser
 if (process.argv[3] == '--min') {
+    console.log('browser');
     require('browser-env')();
     require('../dist/bitdepth.min.js');
-    bitdepth = window.bitdepth.bitdepth;
+    bitdepth = window.bitdepth;
 
 // UMD
 } else if (process.argv[3] == '--umd') {
-	bitdepth = require('../dist/bitdepth.umd.js').bitdepth;
+	console.log('umd');
+	bitdepth = require('../dist/bitdepth.umd.js');
 
 // CommonJS
 } else if (process.argv[3] == '--cjs') {
-	bitdepth = require('../dist/bitdepth.cjs.js').bitdepth;
+	console.log('cjs');
+	bitdepth = require('../dist/bitdepth.cjs.js');
+
+// esm
+} else if (process.argv[3] == '--esm') {
+	console.log('esm');
+	bitdepth = require('../dist/bitdepth.js').default;
 
 // ESM
 } else {
-	bitdepth = require('../index.js').bitdepth;
+	console.log('source');
+	bitdepth = require('../index.js').default;
 }
 
 module.exports = bitdepth;