+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v6.0.0 (2018-06-25)
+- Allow better use of this lib as a dependency:
+	- package.json refactored with bundlers and ES6 envs in mind
+	- bitdetph function exported as default
+
 ## 5.0.0 (2018-06-22)
 - ES6 module.